proteinflex/models/analysis/drug_binding.py
from proteinflex.models.utils.lazy_imports import numpy, torch, openmm
# MIT License
#
# Copyright (c) 2024 VishwamAI
#
# Permission is hereby granted, free of charge, to any person obtaining a copy
# of this software and associated documentation files (the "Software"), to deal
# in the Software without restriction, including without limitation the rights
# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
# copies of the Software, and to permit persons to whom the Software is
# furnished to do so, subject to the following conditions:
#
# The above copyright notice and this permission notice shall be included in all
# copies or substantial portions of the Software.
#
# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
# SOFTWARE.

"""Helper module for drug binding analysis"""
from typing import List, Dict, Optional, Tuple
from Bio.PDB import *
from rdkit import Chem
from rdkit.Chem import AllChem
import logging
from openmm import unit
from proteinflex.models.utils.openmm_utils import setup_simulation, minimize_and_equilibrate

logger = logging.getLogger(__name__)

class DrugBindingAnalyzer:
    def __init__(self, esm_model, device):
        """Initialize DrugBindingAnalyzer with ESM model and device"""
        self.esm_model = esm_model
        self.device = device
        self.force_field = 'amber14-all.xml'
        self.temperature = 300  # Kelvin

    def analyze_binding_sites(self, sequence: str) -> List[Dict]:
        """Identify and analyze potential binding sites"""
        try:
            # Get sequence embeddings and attention
            embeddings = self._get_embeddings(sequence)

            # Identify potential binding pockets
            binding_sites = self._identify_pockets(embeddings)

            # Analyze pocket properties
            for site in binding_sites:
                site['properties'] = self._analyze_pocket_properties(sequence, site)

            return binding_sites
        except Exception as e:
            return []

    def _get_embeddings(self, sequence: str) -> torch.Tensor:
        """Get ESM embeddings for sequence"""
        # Implementation
        return torch.zeros(1)  # Placeholder

    def _identify_pockets(self, embeddings: torch.Tensor) -> List[Dict]:
        """Identify potential binding pockets"""
        # Implementation
        return []  # Placeholder

    def _analyze_pocket_properties(self, sequence: str, pocket: Dict) -> Dict:
        """Analyze properties of a binding pocket"""
        # Implementation
        return {}  # Placeholder
