tests/test_drug_binding.py
import pytest
import torch
import numpy as np
from unittest.mock import Mock, patch
from rdkit import Chem
from Bio.PDB import *
from proteinflex.models.analysis.drug_binding import DrugBindingAnalyzer
from proteinflex.models.utils.openmm_utils import setup_simulation
from tests.utils.mock_objects import MockOpenMMSimulation, MockRDKitMol, MockQuantity

@pytest.fixture
def mock_esm_model():
    model = Mock()
    model.forward = Mock(return_value=(
        torch.randn(1, 10, 768),  # Mock embeddings
        torch.randn(1, 8, 10, 10)  # Mock attention maps
    ))
    return model

@pytest.fixture
def mock_device():
    return torch.device('cpu')

@pytest.fixture
def mock_ligand():
    mol = MockRDKitMol()
    return mol

@pytest.fixture
def mock_simulation():
    return MockOpenMMSimulation()

@pytest.fixture
def drug_binding_analyzer(mock_esm_model, mock_device):
    return DrugBindingAnalyzer(mock_esm_model, mock_device)

@pytest.fixture
def sample_sequence():
    return "MAEGEITTFTALTEKFNLPPGNYKKPKLLYCSNG"

def test_init(mock_esm_model, mock_device):
    """Test initialization of DrugBindingAnalyzer"""
    analyzer = DrugBindingAnalyzer(mock_esm_model, mock_device)
    assert analyzer.esm_model == mock_esm_model
    assert analyzer.device == mock_device
    assert analyzer.force_field == 'amber14-all.xml'
    assert analyzer.temperature == 300

def test_prepare_ligand(drug_binding_analyzer, mock_ligand):
    """Test ligand preparation"""
    with patch('rdkit.Chem.AddHs', return_value=mock_ligand), \
         patch('rdkit.Chem.AllChem.EmbedMolecule'), \
         patch('rdkit.Chem.AllChem.MMFFOptimizeMolecule'):
        prepared_ligand = drug_binding_analyzer.prepare_ligand(mock_ligand)
        assert prepared_ligand is not None

def test_prepare_protein(drug_binding_analyzer, sample_sequence):
    """Test protein preparation from sequence"""
    structure = drug_binding_analyzer.prepare_protein_from_sequence(sample_sequence)
    assert isinstance(structure, Structure.Structure)
    assert len(list(structure.get_chains())) == 1

def test_calculate_binding_energy(drug_binding_analyzer, mock_ligand, sample_sequence, mock_simulation):
    """Test binding energy calculation"""
    with patch('proteinflex.models.utils.openmm_utils.setup_simulation', return_value=mock_simulation), \
         patch('proteinflex.models.utils.openmm_utils.minimize_and_equilibrate') as mock_minimize:

        mock_minimize.return_value = mock_simulation.getState()
        result = drug_binding_analyzer.calculate_binding_energy(mock_ligand, sample_sequence)

        assert isinstance(result, dict)
        assert 'binding_energy' in result
        assert 'pose_rmsd' in result
        assert 'success' in result
        assert result['success'] is True

def test_calculate_binding_energy_error_handling(drug_binding_analyzer):
    """Test error handling in binding energy calculation"""
    # Test with None ligand
    with pytest.raises(ValueError):
        drug_binding_analyzer.calculate_binding_energy(None, "SEQUENCE")

    # Test with empty sequence
    with pytest.raises(ValueError):
        drug_binding_analyzer.calculate_binding_energy(MockRDKitMol(), "")

def test_calculate_rmsd(drug_binding_analyzer, mock_ligand, mock_simulation):
    """Test RMSD calculation"""
    state = mock_simulation.getState()
    rmsd = drug_binding_analyzer._calculate_rmsd(mock_ligand, state)
    assert isinstance(rmsd, float)
    assert rmsd >= 0

def test_analyze_binding_sites_success(drug_binding_analyzer, sample_sequence):
    """Test successful binding site analysis"""
    with patch.object(drug_binding_analyzer, '_get_embeddings') as mock_get_embeddings, \
         patch.object(drug_binding_analyzer, '_identify_pockets') as mock_identify_pockets, \
         patch.object(drug_binding_analyzer, '_analyze_pocket_properties') as mock_analyze_properties:

        # Setup mocks
        mock_get_embeddings.return_value = torch.randn(1, 10, 768)
        mock_identify_pockets.return_value = [{'start': 0, 'end': 5}]
        mock_analyze_properties.return_value = {'hydrophobicity': 0.5}

        # Run analysis
        binding_sites = drug_binding_analyzer.analyze_binding_sites(sample_sequence)

        # Verify results
        assert isinstance(binding_sites, list)
        assert len(binding_sites) > 0
        assert 'properties' in binding_sites[0]

        # Verify mock calls
        mock_get_embeddings.assert_called_once_with(sample_sequence)
        mock_identify_pockets.assert_called_once()
        mock_analyze_properties.assert_called_once()

def test_analyze_binding_sites_error_handling(drug_binding_analyzer):
    """Test error handling in binding site analysis"""
    # Test with None sequence
    result = drug_binding_analyzer.analyze_binding_sites(None)
    assert isinstance(result, list)
    assert len(result) == 0

    # Test with empty sequence
    result = drug_binding_analyzer.analyze_binding_sites("")
    assert isinstance(result, list)
    assert len(result) == 0

def test_analyze_binding_sites_integration(drug_binding_analyzer, sample_sequence):
    """Test full binding site analysis pipeline"""
    # Run full analysis
    binding_sites = drug_binding_analyzer.analyze_binding_sites(sample_sequence)

    # Verify basic structure and types
    assert isinstance(binding_sites, list)
    # Even with placeholder implementations, should return empty list
    assert isinstance(binding_sites, list)

@pytest.mark.parametrize("sequence", [
    "MAEGEITTFTALTEKFNLPPGNYKKPKLLYCSNG",  # Normal sequence
    "A" * 100,  # Long sequence
    "M",        # Single amino acid
])
def test_analyze_binding_sites_different_sequences(drug_binding_analyzer, sequence):
    """Test binding site analysis with different sequence lengths"""
    binding_sites = drug_binding_analyzer.analyze_binding_sites(sequence)
    assert isinstance(binding_sites, list)
